Fix auth sample in minimal APIs /3
