+++ aspnetcore/fundamentals/minimal-apis.md
@@ -690,7 +690,7 @@ We recommend adding an extension method to <xref:Microsoft.AspNetCore.Http.IResu
 
 Routes can be protected using authorization policies. These can be declared via the [`[Authorize]`](xref:Microsoft.AspNetCore.Authorization.AuthorizeAttribute) attribute or by using the <xref:Microsoft.AspNetCore.Builder.AuthorizationEndpointConventionBuilderExtensions.RequireAuthorization%2A> method.
 
-[!code-csharp[](minimal-apis/samples/WebRPauth/Program.cs?name=snippet_auth1)]
+[!code-csharp[](minimal-apis/samples/WebRPauth/Program.cs?name=snippet_auth1?highlight=7-8,22)]
 
 The preceding code can be written with <xref:Microsoft.AspNetCore.Builder.AuthorizationEndpointConventionBuilderExtensions.RequireAuthorization%2A>:
 
@@ -698,7 +698,7 @@ The preceding code can be written with <xref:Microsoft.AspNetCore.Builder.Author
 
 The following sample uses [policy-based authorization](xref:security/authorization/policies):
 
-[!code-csharp[](minimal-apis/samples/WebMinAPIs/Program.cs?name=snippet_auth3)]
+[!code-csharp[](minimal-apis/samples/WebRPauth/Program.cs?name=snippet_auth3)]
 
 ### Allow unauthenticated users to access an endpoint
