+++ aspnetcore/fundamentals/minimal-apis/samples/WebRPauth/Program.cs
@@ -1,4 +1,4 @@
-#define FIRST
+#define AUTH3 // FIRST AUTH3
 
 #if DEFAULT
 using Microsoft.AspNetCore.Identity;
@@ -66,6 +66,40 @@ var app = builder.Build();
 app.UseAuthorization();
 
 app.MapGet("/auth", [Authorize] () => "This endpoint requires authorization.");
+app.MapGet("/", () => "This endpoint doesn't require authorization.");
+app.MapGet("/Identity/Account/Login", () => "Sign in page at this endpoint.");
+
+app.Run();
+#endregion
+#elif AUTH3
+#region snippet_auth3
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using WebRPauth.Data;
+
+var builder = WebApplication.CreateBuilder(args);
+builder.Services.AddAuthorization(o => o.AddPolicy("AdminsOnly", 
+                                  b => b.RequireClaim("admin", "true")));
+
+var connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
+builder.Services.AddDbContext<ApplicationDbContext>(options =>
+    options.UseSqlServer(connectionString));
+builder.Services.AddDatabaseDeveloperPageExceptionFilter();
+
+builder.Services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.RequireConfirmedAccount = true)
+    .AddEntityFrameworkStores<ApplicationDbContext>();
+
+var app = builder.Build();
+
+app.UseAuthorization();
+
+app.MapGet("/admin", [Authorize("AdminsOnly")] () => 
+                             "The /admin endpoint is for admins only.");
+
+app.MapGet("/admin2", () => "The /admin2 endpoint is for admins only.")
+   .RequireAuthorization("AdminsOnly");
+
 app.MapGet("/", () => "This endpoint doesn't require authorization.");
 app.MapGet("/Identity/Account/Login", () => "Sign in page at this endpoint.");
